ref: move payments into Landing directory
