+++ frontend/src/pages/Landing/Home/LandingPage.tsx
@@ -50,6 +50,17 @@ import { MarkdownText } from '~components/MarkdownText'
 import { Tab } from '~components/Tabs'
 import { LottieAnimation } from '~templates/LottieAnimation'
 
+import { ExternalFormLink } from '../components/ExternalFormLink'
+import { FeatureGridItem } from '../components/FeatureGridItem'
+import { FeatureLink } from '../components/FeatureLink'
+import { FeatureSection } from '../components/FeatureSection'
+import { HelpAccordionItem } from '../components/HelpAccordionItem'
+import { LandingSection } from '../components/LandingSection'
+import { OrderedListIcon } from '../components/OrderedListIcon'
+import { SectionBodyText } from '../components/SectionBodyText'
+import { SectionTitleText } from '../components/SectionTitleText'
+import { StatsItem } from '../components/StatsItem'
+
 import formsHeroAnimation from './assets/images/animation-hero.json'
 import howFormsWorksAnimation from './assets/images/animation-mode.json'
 import enterEmailAnimation from './assets/images/animation-typing.json'
@@ -72,16 +83,6 @@ import restrictedMsfLogo from './assets/images/restricted__msf.png'
 import restrictedNparksLogo from './assets/images/restricted__nparks.png'
 import restrictedPaLogo from './assets/images/restricted__pa.png'
 import storageModeImg from './assets/images/storage_mode.svg'
-import { ExternalFormLink } from './components/ExternalFormLink'
-import { FeatureGridItem } from './components/FeatureGridItem'
-import { FeatureLink } from './components/FeatureLink'
-import { FeatureSection } from './components/FeatureSection'
-import { HelpAccordionItem } from './components/HelpAccordionItem'
-import { LandingSection } from './components/LandingSection'
-import { OrderedListIcon } from './components/OrderedListIcon'
-import { SectionBodyText } from './components/SectionBodyText'
-import { SectionTitleText } from './components/SectionTitleText'
-import { StatsItem } from './components/StatsItem'
 import { useLanding } from './queries'
 
 export const LandingPage = (): JSX.Element => {
