+++ frontend/src/pages/Landing/Home/assets/images/animation-hero.json
