+++ frontend/src/pages/Landing/Home/assets/images/animation-mode.json
