+++ frontend/src/pages/Landing/Home/assets/images/animation-typing.json
