+++ frontend/src/pages/Landing/Payments/LandingPaymentsPage.tsx
@@ -8,10 +8,10 @@ import { GUIDE_PAYMENTS_PUBLIC } from '~constants/links'
 import { useIsMobile } from '~hooks/useIsMobile'
 import Button from '~components/Button'
 
-import { FeatureGridItem } from '../Landing/components/FeatureGridItem'
-import { LandingSection } from '../Landing/components/LandingSection'
-import { SectionBodyText } from '../Landing/components/SectionBodyText'
-import { SectionTitleText } from '../Landing/components/SectionTitleText'
+import { FeatureGridItem } from '../components/FeatureGridItem'
+import { LandingSection } from '../components/LandingSection'
+import { SectionBodyText } from '../components/SectionBodyText'
+import { SectionTitleText } from '../components/SectionTitleText'
 
 import paymentsImg from './assets/images/graphic_payments.svg'
 import featureFlexibleImg from './assets/images/icon_pay_flexible.svg'
